Fixed a bug where GL_ID were being averaged
GMT gmtmath will average over all columns if -Ca is specified,
but this means that site ID's are averaged along with site
properties.  Instead, we remove -Ca (STAT_ID, which should be
the same as GL_ID, is now diverging as it is in one of the
middle columns and so it becomes an average of the
GL_ID for that site while GL_ID, being the first column, is not
averaged since it is the GMT "time column" and so it takes on
the ID of the last site in the average due to the -Sl option).