Added script to cut to 25 inputs, collab merged data
The cut_columns notebook is used to cut the data from the 100+
features to something more manageable.  This process needs to
be repeated more formally with an FPI, but for now this file
will suffice. The resulting .csv files also have 8 additional
collaborator proposed sites (see site IDs less than 100,000 at
the end of the file).