Adjust column headers to match ICON_ModEx_Data (2)
Work with step_10_output.csv - this takes a long time to build,
so do not do that now.  This file is the source for cut_columns.ipynb
which then generates the WH_RA_GL...<ixy|csv> files. step_10_output.csv
here is the closest to the "colocated" files generated by run_colocate
and archived in dynamic-learning-rivers.
TODO: run_colocate on the raw GLORICH data and do a regression comparison
with step_10_output.csv to validate run_colocate.